Resolve bug image preview not displaying when editing product (#484)
* Resolve bug image preview not displaying when editing product

The endpoint to get the image required authentication. A 302 status code was being returned with a redirection to the login page. This was resolved by adding the token to the request headers when getting the image

* Add tests for helpers/dataLoaders/fetchProtectedImage in react-utils

* Add tests

* Resolve ts lint error unexpected any

Disable rule in packages/react-utils/src/helpers/tests/dataLoaders.test.ts

* Remove unnecessary comment

* Translate untranslated file

* Enhance comment

* Revoke created object URL for image when unmounting product form

Revoke on umount for proper memory managment