Move extracting formValues and creating payload to util functions
fixes #716
Issue: keycloak wrapps any attribute's property value in an array, so when editing
antd rule requiring the contact field to be a string fails because the value passed
to initialValues for contact is an array of string.

Fixed by:
Refactor the formFields type to be explicit about what fields the form handles.
Add util functions to help convert payload to formvalues and back
keycloak user id will now be generated form the FE
fix regressions in tests: