Persist form field values across re-renders.
**Problem:**
The newLocation Container uses react-query which makes requests when events such as window re-focus happens,
causing rerenders in this component.
The initialValues of the form are recomputed a new change each time the component rerenders.
- the locationForm component itself updates the initialValues each time they cahnge,
which is everytime there is a re-render I think. This implicitely overrides the form field values state,
 including those that had been filled already.

**Implemented solution:**
- Preserve the current form field state when the initialValues are reset each time they change in the props.