Change the start and build commands
Running into a load of errors, something to do with webpack module resolution
"BREAKING CHANGE: The request failed to resolve only because it was resolved as fully specified (probably because the origin is strict EcmaScript Module, e. g. a module with javascript mimetype, a '.mjs' file, or a '.js' file where the package.json contains '"type": "module"')."

it would be worth noting also we have potentially 2 versions of react and react-scripts
that might not want to co-exist